Show exercises in CurrentWorkoutActivity.kt

+++ data/src/main/java/vzijden/workout/data/dao/EquipmentDao.kt
@@ -7,5 +7,5 @@ import vzijden.workout.data.model.EquipmentPojo
 @Dao
 interface EquipmentDao {
     @Insert
-    fun insert(equipmentPojo: EquipmentPojo): Int
+    fun insert(equipmentPojo: EquipmentPojo): Long
 }
