+++ domain/src/main/java/vzijden/workout/domain/model/Exercise.kt
@@ -4,10 +4,11 @@ data class Exercise(
     var name: String,
     var description: String
 ) {
-  constructor(name: String, description: String, id: Int, muscleGroups: List<MuscleGroup>): this(name, description) {
+  constructor(name: String, description: String, id: Long, muscleGroups: List<MuscleGroup>) : this(name, description) {
     this.id = id
     this.muscleGroups = muscleGroups
   }
-  var id: Int = 0
+
+  var id: Long = 0
   var muscleGroups: List<MuscleGroup> = mutableListOf()
 }
